+++ zcash_keys/CHANGELOG.md
@@ -13,7 +13,7 @@ and this library adheres to Rust's notion of
 - `impl std::error::Error for DerivationError`
 
 ### Changed
-- Migrated to `orchard 0.10`, `sapling-crypto 0.3`.
+- Migrated to `orchard 0.10`, `sapling-crypto 0.3`, `zcash_protocol 0.4`.
 
 ## [0.3.0] - 2024-08-19
 ### Notable changes